+++ ux/src/UX-Export-Anywhere-Script.js
@@ -1,0 +1,163 @@
+const exportButton = '#exportAnywhere';
+const exportTag = '[Plex EX Export Anywhere]';
+function log(message) {
+    console.log(`${exportTag}[${getTimestamp()}] ${message}`);
+  }
+
+function debug(message) {
+    console.debug(`${exportTag}[${getTimestamp()}] ${message}`);
+  }
+function asDoubleDigit(value) {
+    return value < 10 ? '0' + value : value;
+  }
+
+function getTimestamp() {
+    let dt = new Date();
+    let time = asDoubleDigit(dt.getHours()) + ':' + asDoubleDigit(dt.getMinutes()) + ':' + asDoubleDigit(dt.getSeconds());
+    return time;
+  }
+
+function exportWrapper(){
+    if (document.getElementsByClassName('plex-actions').length == 0){
+        return
+    }
+    exportButtonCreate();
+    addExportCheckboxesToTableHeader();
+}
+
+function generateExportTableObject(){
+    var tbl = $(jQuery('div.plex-grid-wrapper > table')[0]);
+    debug(tbl)
+    var columnIndices = getCheckboxStatus();
+    
+
+    var allColumns = false; 
+    // Fallback to download all columns if there are no checked columns or there is an issue finding the checked columns.
+    if (columnIndices.length < 1){
+        allColumns = true;
+    }
+
+    var rowDataList = [];
+    var headers = [];
+
+    // Getting the text from just the th element includes some hidden text content from the sort order of the column.
+    // Some columns don't have text and this causes the indexes to be desynced if trying to directly find abbr elements.
+    // Need to include the tr class because there is a single phantom thead tr th element for some reason
+    tbl.find('thead tr.plex-grid-header-row th').each(function(index) {
+        var columnText = ''
+        if (this.querySelector('div abbr')){
+            columnText = $(this.querySelector('div abbr')).text().trim();
+            debug('Column has abbr tag')
+        } else {
+            columnText = 'column' + (index)
+            debug('Column does not have abbr tag')
+        }
+        debug(`Column text: ${columnText}`)
+        headers[index] = columnText
+    });
+
+    if (allColumns){
+        columnIndices = Array.from({length: headers.length}, (_, index) => index);
+    }
+
+    tbl.find('tbody tr').each(function() {
+        debug('compiling row data')
+        
+        var rowData = {}; 
+        var cells = $(this).find('td'); 
+
+        // Loop through the specified column indices and store the values in the object
+        columnIndices.forEach(function(index) {
+            if (cells.eq(index).length) {
+                // Use the header text as the key and the cell content as the value
+                var headerText = headers[index] || ('column' + (index)); // Fallback to 'columnX' if no header. 
+                rowData[headerText] = cells.eq(index).text().trim().replaceAll('"','""'); //Not having this will break csv formatting.
+            }
+        });
+
+        // Add the object to the list if it has any data
+        if (Object.keys(rowData).length > 0) {
+            rowDataList.push(rowData);
+            debug(rowData)
+        }
+    });
+
+    debug('Extracted row data with headers as keys:', rowDataList);
+    return rowDataList
+}
+
+function saveExcportJsonToCsv() {
+    const jsonObject = generateExportTableObject()
+    const csv = convertExportJsonToCsv(jsonObject);
+    const saveTitle = document.title
+    exportCsv(csv, `${saveTitle}.csv`);
+}
+
+function convertExportJsonToCsv(jsonData) {
+    const headers = Object.keys(jsonData[0]);
+    const csvRows = jsonData.map(row => 
+        headers.map(header => `"${row[header]}"`).join(',')
+    );
+
+    csvRows.unshift(headers.join(','));
+
+    return csvRows.join('\n');
+}
+
+function exportCsv(csvContent, fileName) {
+    const blob = new Blob([csvContent], { type: 'text/csv' });
+    const url = window.URL.createObjectURL(blob);
+
+    const a = document.createElement('a');
+    a.setAttribute('href', url);
+    a.setAttribute('download', fileName);
+    a.click();
+    
+    window.URL.revokeObjectURL(url);
+}
+
+function exportButtonCreate(){
+    uxCreateButton('exportAnywhere','Export Visible Content')
+    document.getElementById('exportAnywhere').addEventListener('click', saveExcportJsonToCsv);
+}
+function getCheckboxStatus(){
+    const selectedColumns = document.querySelectorAll('.PlexEX.ExportAnywhereCheckbox')
+    let selectedIndexes = []
+    for (let i = 0; i < selectedColumns.length; i++){
+        if (selectedColumns[i].checked){
+            selectedIndexes.push(i)
+        }
+    }
+    return selectedIndexes
+}
+async function addExportCheckboxesToTableHeader() {
+    try {
+        const table = await waitForElement(exportButton, 60000, document, 0)
+        var tbl = jQuery('div.plex-grid-wrapper > table')[0]
+        debug(table)
+        const headerRow = await waitForElement('thead tr.plex-grid-header-row', 60000, tbl, 0);
+        debug(headerRow)
+        const headers = headerRow.getElementsByTagName('th');
+        debug(headers)
+        for (let i = 0; i < headers.length; i++) {
+            const checkbox = document.createElement('input');
+            checkbox.type = 'checkbox';
+            checkbox.checked = true;
+            checkbox.setAttribute('class', 'PlexEX ExportAnywhereCheckbox')
+            
+            const checkboxContainer = document.createElement('div');
+            checkboxContainer.style.textAlign = 'center';
+            
+            checkboxContainer.appendChild(checkbox);
+            
+            // Placing the checkbox under the column text is more consistently formatted.
+            // TODO Want to more elegantly handle the checkbox placement.
+            // The layout is bad for columns that are themselves a checkbox.
+            // Want to see if there can be a DIV or SPAN inserted to align them.
+            headers[i].appendChild(checkboxContainer);
+        }
+    } catch (error){
+        console.error(error);
+    }
+}
+exportWrapper();